CODA-204 - Rename package domain
