+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.androidwebapp"
+        applicationId "net.coda_it.androidwebapp"
         minSdk 21
         targetSdk 31
         versionCode 1
