--- app/src/main/java/com/example/androidwebapp/domain/models/post/PostModel.java
@@ -1,10 +1,0 @@
-package com.example.androidwebapp.domain.models.post;
-
-public class PostModel {
-    public String id;
-    public String userId;
-    public String categoryId;
-    public String title;
-    public String description;
-    public String image;
-}
