--- app/src/main/java/com/example/androidwebapp/domain/models/user/UserModel.java
@@ -1,4 +1,0 @@
-package com.example.androidwebapp.domain.models.user;
-
-public class UserModel {
-}
