+++ app/src/main/java/net/coda_it/androidwebapp/domain/models/post/PostModel.java
@@ -1,0 +1,10 @@
+package net.coda_it.androidwebapp.domain.models.post;
+
+public class PostModel {
+    public String id;
+    public String userId;
+    public String categoryId;
+    public String title;
+    public String description;
+    public String image;
+}
