+++ app/src/main/java/net/coda_it/androidwebapp/domain/models/user/UserModel.java
@@ -1,0 +1,4 @@
+package net.coda_it.androidwebapp.domain.models.user;
+
+public class UserModel {
+}
