+++ app/src/main/java/net/coda_it/androidwebapp/domain/usecases/user/UserUsecases.java
@@ -1,0 +1,17 @@
+package net.coda_it.androidwebapp.domain.usecases.user;
+
+import android.app.Activity;
+
+import net.coda_it.androidwebapp.data.repositories.user.UserRepository;
+
+public class UserUsecases {
+    UserRepository userRepository;
+
+    public UserUsecases(Activity activity) {
+        this.userRepository = new UserRepository(activity);
+    }
+
+    public void login(String login, String password) {
+        this.userRepository.login(login, password);
+    }
+}
